export_onnx/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
diffusers==0.24.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
fire==0.7.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
grpcio==1.66.1
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lap==0.4.0
lazy_loader==0.4
lit==18.1.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
onnx==1.16.0
onnxruntime-gpu==1.17.1
onnxsim==0.4.36
onnxslim @ file:///root/github_for_install/onnxslim-0.1.34-py3-none-any.whl#sha256=755cb13cfd7a0b47d33e5c1935db2bb7a6993e9098a33987c28794f9187cab17
opencv-python==4.7.0.72
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybboxes==0.1.6
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
safetensors==0.4.5
sahi==0.11.20
scikit-image==0.21.0
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
stack-data==0.6.3
sympy==1.13.3
tb-nightly==2.14.0a20230808
tensorboard-data-server==0.7.2
tensorrt @ file:///root/TensorRT-8.6.1.6/python/tensorrt-8.6.1-cp310-none-linux_x86_64.whl#sha256=5cfdc04622fa824b937d445ccd00aee04cffcd21482b000d184ca27059957fd4
termcolor==2.5.0
terminaltables==3.1.10
tifffile==2024.8.30
torch==2.0.1
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
triton==2.0.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics-thop==2.0.6
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.2
